@@ -17,9 +17,8 @@ classifiers =
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
@@ -30,13 +29,10 @@ classifiers =
 
 [options]
 packages = redis
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*
+python_requires = >=3.5
 
 [options.extras_require]
 hiredis = hiredis>=0.1.3
 
 [flake8]
 exclude = .venv,.tox,dist,docs,build,*.egg,redis_install,env,venv,.undodir
-
-[bdist_wheel]
-universal = 1
