@@ -3,10 +3,9 @@ import itertools
 import redis
 import sys
 import timeit
-from redis._compat import izip
 
 
-class Benchmark(object):
+class Benchmark:
     ARGUMENTS = ()
 
     def __init__(self):
@@ -34,7 +33,7 @@ class Benchmark(object):
         group_names = [group['name'] for group in self.ARGUMENTS]
         group_values = [group['values'] for group in self.ARGUMENTS]
         for value_set in itertools.product(*group_values):
-            pairs = list(izip(group_names, value_set))
+            pairs = list(zip(group_names, value_set))
             arg_string = ', '.join(['%s=%s' % (p[0], p[1]) for p in pairs])
             sys.stdout.write('Benchmark: %s... ' % arg_string)
             sys.stdout.flush()
